// src/utils/enums.ts
export enum TextVariants {
  INTER_600_32_40 = 'INTER_600_32_40',
  INTER_600_18_24 = 'INTER_600_18_24',
  INTER_600_16_26 = 'INTER_600_16_26',
  INTER_600_14_20 = 'INTER_600_14_20',
  INTER_600_12_18 = 'INTER_600_12_18',
  INTER_500_20_18 = 'INTER_500_20_18',
  INTER_500_16_44 = 'INTER_500_16_44',
  INTER_500_14_44 = 'INTER_500_14_44',
  INTER_500_12_18 = 'INTER_500_12_18',
  INTER_400_14_20 = 'INTER_400_14_20',
  INTER_400_12_18 = 'INTER_400_12_18',
}